requirements.txt
Flask==3.0.0
pymongo==4.6.0
flask-cors==4.0.0
Werkzeug==3.0.1
razorpay==2.0.0
python-dotenv==1.0.0
requests==2.31.0
apscheduler==3.10.4
google-generativeai==0.3.1
Flask-Mail==0.9.1
Authlib==1.3.0
